requirements.txt
torch>=2.2
torchvision>=0.17
transformers>=4.41
datasets>=2.18
pandas>=2.2
scikit-learn>=1.4
tqdm>=4.66
emoji>=2.11
Pillow>=10.0
